Cleanup and fixes around tracked position mode
Crude velocity damping on tracked position targets reduces jerk on move starts etc

+++ firmware/src/drivers/motion_types.h
@@ -114,7 +114,7 @@ cartesian_find_point_on_line( CartesianPoint_t *a, CartesianPoint_t *b, Cartesia
 PUBLIC int32_t
 cartesian_distance_between( CartesianPoint_t *a, CartesianPoint_t *b );
 
-PUBLIC int16_t
+PUBLIC uint32_t
 cartesian_duration_for_speed( CartesianPoint_t *a, CartesianPoint_t *b, mm_per_second_t target_speed );
 
 PUBLIC KinematicsSolution_t
